Update projects.html
Deleted the completed projects section of the table as the information is no longer relevant nor will it ever be updated. However the information found within the discarded links can still be found in the projects folder in the page section.  Also the table has changed to a complete 4 column table instead of a part 4 column and part 5 column table as it was before.  Pictures will be later added in the caption above each link as well as in the projects pages themselves (see projects folder for more information on each new completed project)  Uncertain if the last line in the page is still relevant.